Renames user-reducer and redux state from user to account (makes more sense, aligns with the actions, and is easier to find
