+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/account-menu.js
@@ -75,8 +75,8 @@ function genLogoutConf() {
       this.maxEmailLength = 16;
 
       const logout = state => ({
-        loggedIn: state.getIn(["user", "loggedIn"]),
-        email: state.getIn(["user", "email"]),
+        loggedIn: state.getIn(["account", "loggedIn"]),
+        email: state.getIn(["account", "email"]),
         isPrivateIdUser: isPrivateUser(state),
       });
